[Fix] Fix build error
